add: コンテンツプロセッサーの型定義とディレクトリローダーを実装
