+++ packages/content-processor/src/loaders/load-directory.ts
@@ -1,7 +1,7 @@
 import type { DirectoryLoaderResult } from '../types';
 
 // 仮実装: loadDirectory は配列を返す
-export function loadDirectory(path: string): DirectoryLoaderResult {
-  // TODO: 実装は後で追加
+export function loadDirectory(options: Record<string, unknown>): DirectoryLoaderResult {
+  // 仮実装: 空配列を返す
   return [];
 }
