--- packages/content-processor/src/types/core/loader.ts
@@ -1,2 +1,0 @@
-// 仮定義: DirectoryLoaderResult
-export type DirectoryLoaderResult = any[];
