+++ packages/content-processor/src/types/index.ts
@@ -1,10 +1,2 @@
-// コア型のエクスポート
-export type { PostMeta, PostHTML } from './core/post';
-export type { ProcessorOptions, ListOptions } from './core/options';
-export type { DirectoryLoaderResult } from './core/loader';
-
-// エラークラスのエクスポート
-export { FileNotFoundError, FrontMatterError, MarkdownParseError } from '../errors';
-
-// 型ユーティリティ
-export type { Schema } from 'hast-util-sanitize';
+export * from './post';
+export * from './options';
